refactor: HomePage InvitationPage 이름변경
