+++ src/App.tsx
@@ -4,7 +4,7 @@ import { Helmet } from 'react-helmet-async'
 import { Switch, Route, BrowserRouter } from 'react-router-dom'
 
 /* Internal dependencies */
-import { HomePage, ErrorPage, LandingPage } from 'pages'
+import { HomePage, ErrorPage, InvitationPage } from 'pages'
 import ErrorHandler from 'components/ErrorHandler'
 import { GlobalStyle } from 'styles/global-styles'
 import 'app.scss'
@@ -17,8 +17,8 @@ function App() {
       </Helmet>
       <ErrorHandler>
         <Switch>
-          <Route exact path="/" component={LandingPage} />
-          <Route exact path="/:templateId" component={HomePage} />
+          <Route exact path="/" component={HomePage} />
+          <Route exact path="/:templateId" component={InvitationPage} />
           <Route component={ErrorPage} />
         </Switch>
       </ErrorHandler>
