+++ src/components/Home/Home.module.scss
@@ -16,7 +16,7 @@
 }
 
 
-.landing {
+.home {
     width: 100%;
     height: auto;
