+++ src/pages/HomePage/HomePage.tsx
@@ -1,36 +1,18 @@
 /* External dependencies */
 import React from 'react'
-import { RouteComponentProps } from 'react-router-dom'
-import { Helmet } from 'react-helmet-async'
 import classNames from 'classnames/bind'
 
 /* Internal dependencies */
-import PopAnimation from 'components/PopAnimation'
-import InvitationContainer from 'containers/InvitationContainer'
-import styles from './HomePage.module.scss'
+import style from 'pages/HomePage/HomePage.module.scss'
+import Home from 'components/Home/Home'
 
-interface RouterProps {
-  templateId: string
-}
+const cx = classNames.bind(style)
 
-const cx = classNames.bind(styles)
-
-function HomePage({ match }: RouteComponentProps<RouterProps>) {
-  const { templateId } = match.params
-
+function HomePage() {
   return (
-    <>
-      <Helmet>
-        <title>Invitation</title>
-        <meta name="description" content="Nawa invitation" />
-      </Helmet>
-      <div className={cx('homepage-wrapper')}>
-        <div className={cx('main-wrapper')}>
-          <PopAnimation duration={1.5} />
-          <InvitationContainer templateId={templateId} />
-        </div>
-      </div>
-    </>
+    <div className={cx('homepage-wrapper')}>
+      <Home />
+    </div>
   )
 }
