--- src/pages/LandingPage/LandingPage.module.scss
@@ -1,4 +1,0 @@
-.landing-page {
-    width: 270px;
-    height: auto;
-}
