+++ src/pages/index.ts
@@ -1,3 +1,3 @@
 export { default as HomePage } from './HomePage'
 export { default as ErrorPage } from './ErrorPage'
-export { default as LandingPage } from './LandingPage'
+export { default as InvitationPage } from './InvitationPage'
